Fix non-deterministic > _NodeCursor comparison err
Sometimes (we don't know why), when building an RTree index, we get an error:
`TypeError: '>' not supported between instances of '_NodeCursor' and '_NodeCursor'`

This ugly hack just teaches python to be able to compare instances of _NodeCursor
by always assuming one is greater than the other. Doesn't seem to make a difference 
to correctness or performance of the index.